+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Nomenclature field in visualize form is now correctly handled when the input is a decoded JSON string.
+
 ## [3.6.1](https://github.com/InseeFr/Drama-Queen/releases/tag/3.6.1) - 2026-06-15
 
 ### Fixed